requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
flask==3.1.1
openai==1.14.0
pillow==10.2.0
python-wordpress-xmlrpc==2.3
requests==2.32.3
urllib3==1.26.18
